refactor: publish one job shape to every consumer
The jobs API published `WorkflowJob` and the web app published `Job`.
Both were built from the same `@virtool/data` record and differed in
three ways, two of which were not real: `pingedAt` was carried only by
`WorkflowJob` and read by nothing, and `args` was typed `JsonObject`
there against `Record<string, string>` here while holding the identical
value. Its docstring described `args` as an opaque blob out of a JSONB
column, but `jobs` has no `args` column — the map is recomposed from the
resources that reference the job. That left `finishedAt`, one nullable
date, as the whole of the difference.

Collapse to one `Job` in `contracts/src/jobs.ts`, now a zod schema
because the jobs API parses on the way out and the runtime parses on the
way in; the SPA still parses nothing and imports the inferred type. Drop
`pingedAt`, which the data layer still reads for `/metrics`, and type
`args` as the string map it is.

The harness records the heartbeat on `JobsApiState` rather than on the
job, mirroring a column the service writes and no read shape publishes.
