+++ docs/jobs-api.md
@@ -366,7 +366,7 @@ metadata reads.
 
 `jobs.workflow` is a plain `text` column with no CHECK constraint, so the
 data layer types it `string` while the wire types it as a union. `toJob`
-therefore **parses** the response through the `WorkflowJob` schema rather
+therefore **parses** the response through the `Job` schema rather
 than asserting it into shape, and a row that does not fit is thrown:
 `app.onError` logs it, hands it to Sentry and answers
 `jsonError(500, "Internal server error")`. The message names the job id
@@ -374,7 +374,8 @@ and the failing field paths — never the values, because a message is a
 Sentry title and row content does not belong in one.
 
 Failing the read is not new strictness. `@virtool/workflow`'s client
-parses what it receives with the same schema, so an unrecognised
+parses what it receives with that same schema — one `Job` serves this
+service, the web app and the runtime alike — so an unrecognised
 workflow was already a hard `JobsApiError` — raised at a runner that can
 do nothing about it, on the far side of a retry policy. Parsing here
 puts the failure on the side that owns the data, where it is one Sentry
