+++ packages/contracts/src/jobs.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, it } from "vitest";
 import {
 	fromStoredJobClaim,
 	fromStoredJobStep,
+	Job,
 	JobClaim,
 	JobStep,
 	StoredJobClaim,
@@ -81,5 +82,114 @@ describe("step round trip", () => {
 		expect(fromStoredJobStep(toStoredJobStep(unstarted))).toStrictEqual(
 			unstarted,
 		);
+	});
+});
+
+// JSON has no date type, so what actually crosses is a string either way. These
+// pin that the encoding is unchanged while the parsed type is not.
+describe("Job timestamps", () => {
+	const CREATED_AT = "2026-07-31T16:30:00.000Z";
+
+	function job(overrides: Record<string, unknown> = {}) {
+		return {
+			id: 1,
+			args: {},
+			claim: null,
+			claimedAt: null,
+			createdAt: CREATED_AT,
+			finishedAt: null,
+			progress: 0,
+			state: "pending",
+			steps: null,
+			user: { id: 3, handle: "igboyes" },
+			workflow: "create_sample",
+			...overrides,
+		};
+	}
+
+	it("parses the string the wire carries into a Date", () => {
+		const parsed = Job.parse(job());
+
+		expect(parsed.createdAt).toBeInstanceOf(Date);
+		expect(parsed.createdAt.toISOString()).toBe(CREATED_AT);
+	});
+
+	// The handler hands `Response.json` the `Date` it read out of Postgres, and
+	// `JSON.stringify` calls `Date.prototype.toJSON`. Python reads these bytes,
+	// so they must not have moved.
+	it("encodes back to the same ISO string it arrived as", () => {
+		const encoded = JSON.parse(JSON.stringify(Job.parse(job())));
+
+		expect(encoded.createdAt).toBe(CREATED_AT);
+	});
+
+	it("passes a Date through unchanged, so one schema types both directions", () => {
+		const parsed = Job.parse(job({ createdAt: new Date(CREATED_AT) }));
+
+		expect(parsed.createdAt.toISOString()).toBe(CREATED_AT);
+	});
+
+	it("keeps a null timestamp null rather than coercing it to the epoch", () => {
+		// `new Date(null)` is the epoch, not an invalid date, so a nullable field
+		// that let `coerce` see the null would silently report 1970.
+		expect(Job.parse(job({ finishedAt: null })).finishedAt).toBeNull();
+	});
+
+	// `coerce` runs `new Date(value)`, which answers `Invalid Date` rather than
+	// throwing. Without the refinement this parses and surfaces as `NaN` much
+	// later in a run.
+	it.each(["not-a-date", "", "2026-13-45T99:99:99Z"])(
+		"rejects %j rather than parsing it as an invalid date",
+		(value) => {
+			expect(Job.safeParse(job({ createdAt: value })).success).toBe(false);
+		},
+	);
+});
+
+describe("Job", () => {
+	const base = {
+		id: 1,
+		args: {},
+		claim: null,
+		claimedAt: null,
+		createdAt: "2026-07-31T16:30:00.000Z",
+		finishedAt: null,
+		progress: 0,
+		state: "pending",
+		steps: null,
+		user: { id: 3, handle: "igboyes" },
+		workflow: "create_sample",
+	};
+
+	it("accepts a build_index job on the read path", () => {
+		// `build_index` stays Python-owned and is never handed out at claim time,
+		// but its rows exist and must still parse.
+		expect(Job.safeParse({ ...base, workflow: "build_index" }).success).toBe(
+			true,
+		);
+	});
+
+	it("reads args as a map of stringified ids", () => {
+		// `jobs` has no `args` column: the map is recomposed from the resources
+		// that reference the job, and every value is an id run through `String`.
+		expect(Job.parse({ ...base, args: { sample_id: "4" } }).args).toStrictEqual(
+			{
+				sample_id: "4",
+			},
+		);
+	});
+
+	it("rejects an arg value that is not a string", () => {
+		expect(Job.safeParse({ ...base, args: { sample_id: 4 } }).success).toBe(
+			false,
+		);
+	});
+
+	// One shape serves the SPA and the workflow runtime. A runner reading an
+	// older build's response must not trip over a field it does not name.
+	it("strips an unknown field rather than refusing the response", () => {
+		const parsed = Job.parse({ ...base, pingedAt: "2026-07-31T16:31:00.000Z" });
+
+		expect(parsed).not.toHaveProperty("pingedAt");
 	});
 });
