+++ packages/contracts/src/jobsApi.test.ts
@@ -7,7 +7,6 @@ import {
 	FinalizeSubtractionRequest,
 	JobClaimed,
 	JobFileManifest,
-	WorkflowJob,
 } from "./jobsApi";
 
 const claim: JobClaim = {
@@ -18,69 +17,6 @@ const claim: JobClaim = {
 	runtimeVersion: "2.0.0",
 	workflowVersion: "1.2.3",
 };
-
-// JSON has no date type, so what actually crosses is a string either way. These
-// pin that the encoding is unchanged while the parsed type is not.
-describe("timestamps", () => {
-	const CREATED_AT = "2026-07-31T16:30:00.000Z";
-
-	function job(overrides: Record<string, unknown> = {}) {
-		return {
-			id: 1,
-			args: {},
-			claim: null,
-			claimedAt: null,
-			createdAt: CREATED_AT,
-			pingedAt: null,
-			progress: 0,
-			state: "pending",
-			steps: null,
-			user: { id: 3, handle: "igboyes" },
-			workflow: "create_sample",
-			...overrides,
-		};
-	}
-
-	it("parses the string the wire carries into a Date", () => {
-		const parsed = WorkflowJob.parse(job());
-
-		expect(parsed.createdAt).toBeInstanceOf(Date);
-		expect(parsed.createdAt.toISOString()).toBe(CREATED_AT);
-	});
-
-	// The handler hands `Response.json` the `Date` it read out of Postgres, and
-	// `JSON.stringify` calls `Date.prototype.toJSON`. Python reads these bytes,
-	// so they must not have moved.
-	it("encodes back to the same ISO string it arrived as", () => {
-		const encoded = JSON.parse(JSON.stringify(WorkflowJob.parse(job())));
-
-		expect(encoded.createdAt).toBe(CREATED_AT);
-	});
-
-	it("passes a Date through unchanged, so one schema types both directions", () => {
-		const parsed = WorkflowJob.parse(job({ createdAt: new Date(CREATED_AT) }));
-
-		expect(parsed.createdAt.toISOString()).toBe(CREATED_AT);
-	});
-
-	it("keeps a null timestamp null rather than coercing it to the epoch", () => {
-		// `new Date(null)` is the epoch, not an invalid date, so a nullable field
-		// that let `coerce` see the null would silently report 1970.
-		expect(WorkflowJob.parse(job({ pingedAt: null })).pingedAt).toBeNull();
-	});
-
-	// `coerce` runs `new Date(value)`, which answers `Invalid Date` rather than
-	// throwing. Without the refinement this parses and surfaces as `NaN` much
-	// later in a run.
-	it.each(["not-a-date", "", "2026-13-45T99:99:99Z"])(
-		"rejects %j rather than parsing it as an invalid date",
-		(value) => {
-			expect(WorkflowJob.safeParse(job({ createdAt: value })).success).toBe(
-				false,
-			);
-		},
-	);
-});
 
 describe("wire shapes", () => {
 	it("parses a claim response and keeps the one-time key", () => {
@@ -109,47 +45,6 @@ describe("wire shapes", () => {
 
 		expect(request.runnerId).toBe("runner-1");
 		expect(request.steps).toHaveLength(1);
-	});
-
-	it("accepts a build_index job on the read path", () => {
-		// `build_index` stays Python-owned and is never handed out at claim time,
-		// but its rows exist and must still parse.
-		expect(
-			WorkflowJob.safeParse({
-				id: 1,
-				args: { index_id: 7 },
-				claim: null,
-				claimedAt: null,
-				createdAt: "2026-07-31T16:30:00.000Z",
-				pingedAt: null,
-				progress: 0,
-				state: "pending",
-				steps: null,
-				user: { id: 3, handle: "igboyes" },
-				workflow: "build_index",
-			}).success,
-		).toBe(true);
-	});
-
-	it("treats args as an opaque JSON object", () => {
-		const job = WorkflowJob.parse({
-			id: 1,
-			args: { sample_id: 4, nested: { deep: [1, "two", null] } },
-			claim: null,
-			claimedAt: null,
-			createdAt: "2026-07-31T16:30:00.000Z",
-			pingedAt: null,
-			progress: 0,
-			state: "pending",
-			steps: null,
-			user: { id: 3, handle: "igboyes" },
-			workflow: "create_sample",
-		});
-
-		expect(job.args).toStrictEqual({
-			sample_id: 4,
-			nested: { deep: [1, "two", null] },
-		});
 	});
 });
