+++ packages/workflow/src/client/client.test.ts
@@ -58,7 +58,7 @@ function job(overrides: Record<string, unknown> = {}) {
 		claim: null,
 		claimedAt: null,
 		createdAt: "2026-08-05T00:00:00Z",
-		pingedAt: null,
+		finishedAt: null,
 		progress: 0,
 		state: "running",
 		steps: null,
