+++ packages/workflow/src/testing/jobsApi/routes.ts
@@ -23,11 +23,11 @@ import {
 	FinalizeSampleRequest,
 	FinalizeSubtractionRequest,
 	isJobStateTerminal,
+	type Job,
 	type JobClaimed,
 	type JobPing,
 	type JobStepStarted,
 	RegisterCacheRequest,
-	type WorkflowJob,
 	type WorkflowSample,
 } from "@virtool/contracts";
 import { cacheKey } from "@virtool/storage";
@@ -108,7 +108,7 @@ function parseId(segment: string | undefined): number | null {
  *
  * A copy, so a caller mutating what it read cannot reach back into the fixture.
  */
-function readJob(state: JobsApiState): WorkflowJob {
+function readJob(state: JobsApiState): Job {
 	return {
 		...state.job,
 		steps: state.job.steps?.map((step) => ({ ...step })) ?? null,
@@ -192,7 +192,7 @@ function handleClaim(
 function handlePing(state: JobsApiState): JobsApiResponse {
 	const pingedAt = state.now();
 
-	state.job.pingedAt = pingedAt;
+	state.pingedAt = pingedAt;
 
 	const ping: JobPing = { pingedAt };
